Rename example component file for build
