+++ src/components/Dialog/Dialog.example.tsx
@@ -15,7 +15,7 @@ export type Props = {
   handleClickAway?: () => void;
 };
 
-export const DialogExampleBase = forwardRef<HTMLDialogElement, Props>(
+export const DialogExample = forwardRef<HTMLDialogElement, Props>(
   (
     {
       portalTargetId,
